Move registration form from lms/djangoapps/ucsd_registration_form to existing app openedx/features/ucsd_features
